Created a button that will toggle book read status
Using prototypical inheritance, the button will change this.status from
read to unread or the other way around. I feel like I took a roundabout
way to get it to work, but it works. In the future, look to make sure
that the project has a solid foundation to prevent major revisions
later.